Adding updated files

+++ Navigation/ReplayBuffer.py
@@ -33,9 +33,8 @@ class ReplayBuffer:
         states = torch.from_numpy(np.vstack([e.state for e in mini_batch if e is not None])).float().to(self.device)
         actions = torch.from_numpy(np.vstack([e.action for e in mini_batch if e is not None])).long().to(self.device)
         rewards = torch.from_numpy(np.vstack([e.reward for e in mini_batch if e is not None])).float().to(self.device)
-        next_states = torch.from_numpy(np.vstack([e.next_state for e in mini_batch if e is not None])).float().to(
-            self.device)
-        dones = torch.from_numpy(np.vstack([e.done for e in mini_batch if e is not None])).float().to(self.device)
+        next_states = torch.from_numpy(np.vstack([e.next_state for e in mini_batch if e is not None])).float().to(self.device)
+        dones = torch.from_numpy(np.vstack([e.done for e in mini_batch if e is not None]).astype(np.uint8)).float().to(self.device)
         return states, actions, rewards, next_states, dones
 
     def __len__(self):
